Compile Cydia to arm64 by linking with APT 1.4~b1.
# Conflicts:
#	.gitignore
#	.gitmodules
#	MobileCydia.mm
#	Scripts/Legacy/makefile
#	uikit.sh

+ Added iPhone X launch image

+++ Source/UI/Packages/View Controllers/FileTable.mm
@@ -94,7 +94,7 @@
                 NSString *directory = [stack lastObject];
                 [stack addObject:[file stringByAppendingString:@"/"]];
                 [files_ replaceObjectAtIndex:i withObject:[NSString stringWithFormat:@"%*s%@",
-                                                           ([stack count] - 2) * 3, "",
+                                                           int(([stack count] - 2) * 3), "",
                                                            [file substringFromIndex:[directory length]]
                                                            ]];
             }
